requirements.txt
Django==1.7.1
PyYAML==3.11
SQLAlchemy==0.9.8
dj-database-url==0.3.0
django-jstemplate==1.3.7
django-nose==1.3
django-postgrespool==0.3.0
djangorestframework==3.0.2
djangorestframework-csv==1.3.3
gunicorn==19.1.1
nose==1.3.4
psycopg2==2.5.4
six==1.9.0
whitenoise==1.0.6
wsgiref==0.1.2
